Move intrahost determination algorithm to state.go and update tests
